pyproject.toml
[project]
name = "council"
version = "0.1.0"
description = "Multi-model autonomous research committee"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
council = "council.main:cli"
